eschew imagemagick
Not gonna lie, shelling out to convert(1) was making me fear RCEs,
and I couldn't convince myself that was unreasonable.

Instead use a mix of PIL and gifsicle to do the transformations we need.
PIL is used for file format conversions to GIF, and in the event that it
is needed, to resize input.  gifsicle resizes GIF input because of silly
reasons.